requirements.txt
faster-whisper
TTS
google-search-results
soundfile
librosa
torch
torchaudio
python-dotenv
